requirements.txt
Pillow==10.4.0
pandas==2.2.2